Optional railtie, BodyProxy flush!
