+++ circuitry-middleware.gemspec
@@ -20,6 +20,7 @@ Gem::Specification.new do |spec|
   spec.require_paths = ['lib']
 
   spec.add_dependency 'circuitry', '>= 1.1.0'
+  spec.add_dependency 'rack', '>= 1.4.5'
 
   spec.add_development_dependency 'pry-nav'
   spec.add_development_dependency 'bundler', '~> 1.8'
